moved Mappings out of the resource so its easier to read
